+++ AzureCommunicationUI/AzureCommunicationUI/CallCompositeOptions/CallCompositeOptions.swift
@@ -6,9 +6,12 @@
 import Foundation
 import UIKit
 
+/// User-configurable options for creating CallComposite.
 public struct CallCompositeOptions {
     var themeConfiguration: ThemeConfiguration?
 
+    /// Creates an instance of CallCompositeOptions with related options.
+    /// - Parameter themeConfiguration: ThemeConfiguration for changing color pattern
     public init(themeConfiguration: ThemeConfiguration) {
         self.themeConfiguration = themeConfiguration
     }