+++ AzureCommunicationUI/AzureCommunicationUI/Redux/Reducer/ErrorReducer.swift
@@ -18,12 +18,16 @@ struct ErrorReducer: Reducer {
         switch action {
         case let action as ErrorAction.FatalErrorUpdated:
             error = action.error
-            errorCode = action.errorCode
+            errorCode = action.error.code
             errorCategory = .fatal
         case let action as ErrorAction.CallStateErrorUpdated:
             error = action.error
-            errorCode = action.errorCode
+            errorCode = action.error.code
             errorCategory = .callState
+        case _ as CallingViewLaunched:
+            error = nil
+            errorCode = ""
+            errorCategory = .none
         default:
             return state
         }