+++ AzureCommunicationUI/AzureCommunicationUI/Utilities/StringConstants.swift
@@ -1,0 +1,11 @@
+//
+//  Copyright (c) Microsoft Corporation. All rights reserved.
+//  Licensed under the MIT License.
+//
+
+import Foundation
+
+struct StringConstants {
+    static let defaultEmptyName: String = "Unnamed participant"
+    static let localParticipantNamePostfix: String = "(me)"
+}