+++ AzureCommunicationUI/AzureCommunicationUITests/Presentation/Setup/SetupViewModelTests.swift
@@ -15,6 +15,7 @@ class SetupViewModelTests: XCTestCase {
     private let timeout: TimeInterval = 10.0
 
     override func setUp() {
+        super.setUp()
         storeFactory = StoreFactoryMocking()
         cancellable = CancelBag()